chore: extract duplicated createTestImage from ImageCompressTests and ImageResizeTests to ImageTestUtils
